pyproject.toml
[project]
name = "webstrike-buisness"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.123.5",
    "psycopg[binary,pool]>=3.3.1",
    "psycopg-pool>=3.3.0",
    "uvicorn>=0.38.0",
    "python-dotenv>=1.2.1",
    "argon2-cffi>=25.1.0",
    "pyjwt>=2.10.1",
]
